:sparkles: Better positioning of search palette
